Move persistence logic into separate module file
Move persistence-related code out of `v2::Sender` and `v2::Receiver`
into dedicated `persist.rs` files. This refactor helps declutter the
core API and separates unrelated concerns.

This change addresses a follow-up item from the [PR that introduced
persistence](https://github.com/payjoin/rust-payjoin/pull/552#discussion_r2020220139).
